requirements.txt
discord.py[voice]
asyncpg
bs4
